clean code TrackView
